api/requirements.txt
click==7.1.2
Flask==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
pkg-resources==0.0.0
psycopg2==2.8.5
python-dotenv==0.13.0
Werkzeug==1.0.1
